+++ app/controllers/seasons/best_players_controller.rb
@@ -14,7 +14,7 @@ module Seasons
     private
 
     def find_season
-      @season = Season.active.find_by!(uuid: params[:season_id])
+      @season = Season.in_progress.find_by!(uuid: params[:season_id])
     end
 
     def find_week