+++ app/jobs/injuries/import_job.rb
@@ -5,7 +5,7 @@ module Injuries
     queue_as :default
 
     def perform(import_service: Injuries::ImportService.new)
-      Season.active.each do |season|
+      Season.in_progress.each do |season|
         import_service.call(season: season)
       end
     end