+++ app/models/league.rb
@@ -4,7 +4,7 @@ class League < ApplicationRecord
   include Sportable
 
   has_many :seasons, dependent: :destroy
-  has_one :active_season, -> { Season.active }, class_name: 'Season', foreign_key: :league_id # rubocop: disable Rails/HasManyOrHasOneDependent
+  has_one :active_season, -> { Season.in_progress }, class_name: 'Season', foreign_key: :league_id # rubocop: disable Rails/HasManyOrHasOneDependent
 
   has_many :cups, dependent: :destroy
 end