+++ spec/controllers/fantasy_teams/transfers_controller_spec.rb
@@ -38,6 +38,16 @@ describe FantasyTeams::TransfersController do
 
           expect(response).to render_template :show
         end
+
+        context 'for finishing season' do
+          before { fantasy_team.season.update!(status: Season::FINISHING) }
+
+          it 'renders 404 page' do
+            get :show, params: { fantasy_team_id: fantasy_team.uuid, locale: 'en' }
+
+            expect(response).to render_template 'shared/404'
+          end
+        end
       end
     end